requirements.txt
Django==1.7.7
django-grappelli==2.6.5
mod-wsgi==4.4.11
mysql-connector-python==2.0.2
sorl-thumbnail==12.2
Pillow==2.8.2
